requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
openai==1.6.1
httpx==0.25.2
python-dotenv==1.0.0
requests==2.31.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Real-time communication
websockets==12.0
asyncio-mqtt==0.16.1

# Development dependencies (optional but recommended)
pytest==7.4.0
pytest-asyncio==0.21.0
black==24.10.0
isort==5.12.0
flake8==6.0.0
mypy==1.5.0 